src/main/webapp/app/entities/chips/route/chips-routing-resolve.service.spec.ts
jest.mock('@angular/router');

import { TestBed } from '@angular/core/testing';
import { HttpResponse } from '@angular/common/http';
import { HttpClientTestingModule } from '@angular/common/http/testing';
import { ActivatedRouteSnapshot, Router } from '@angular/router';
import { of } from 'rxjs';

import { IChips, Chips } from '../chips.model';
import { ChipsService } from '../service/chips.service';

import { ChipsRoutingResolveService } from './chips-routing-resolve.service';

describe('Chips routing resolve service', () => {
  let mockRouter: Router;
  let mockActivatedRouteSnapshot: ActivatedRouteSnapshot;
  let routingResolveService: ChipsRoutingResolveService;
  let service: ChipsService;
  let resultChips: IChips | undefined;

  beforeEach(() => {
    TestBed.configureTestingModule({
      imports: [HttpClientTestingModule],
      providers: [Router, ActivatedRouteSnapshot],
    });
    mockRouter = TestBed.inject(Router);
    mockActivatedRouteSnapshot = TestBed.inject(ActivatedRouteSnapshot);
    routingResolveService = TestBed.inject(ChipsRoutingResolveService);
    service = TestBed.inject(ChipsService);
    resultChips = undefined;
  });

  describe('resolve', () => {
    it('should return IChips returned by find', () => {
      // GIVEN
      service.find = jest.fn(id => of(new HttpResponse({ body: { id } })));
      mockActivatedRouteSnapshot.params = { id: 123 };

      // WHEN
      routingResolveService.resolve(mockActivatedRouteSnapshot).subscribe(result => {
        resultChips = result;
      });

      // THEN
      expect(service.find).toBeCalledWith(123);
      expect(resultChips).toEqual({ id: 123 });
    });

    it('should return new IChips if id is not provided', () => {
      // GIVEN
      service.find = jest.fn();
      mockActivatedRouteSnapshot.params = {};

      // WHEN
      routingResolveService.resolve(mockActivatedRouteSnapshot).subscribe(result => {
        resultChips = result;
      });

      // THEN
      expect(service.find).not.toBeCalled();
      expect(resultChips).toEqual(new Chips());
    });

    it('should route to 404 page if data not found in server', () => {
      // GIVEN
      jest.spyOn(service, 'find').mockReturnValue(of(new HttpResponse({ body: null as unknown as Chips })));
      mockActivatedRouteSnapshot.params = { id: 123 };

      // WHEN
      routingResolveService.resolve(mockActivatedRouteSnapshot).subscribe(result => {
        resultChips = result;
      });

      // THEN
      expect(service.find).toBeCalledWith(123);
      expect(resultChips).toEqual(undefined);
      expect(mockRouter.navigate).toHaveBeenCalledWith(['404']);
    });
  });
});
